Refactor GlobalRevitUiContext to its own source file
Move GlobalRevitUiContext implementation from GlobalRevitContext.cs to GlobalRevitUiContext.cs. Improve documentation, ensure null and validity checks, and clarify API usage for singleton Revit add-ins. Remove old implementation from GlobalRevitContext.cs.
