+++ Source/Scotec.Revit.Test.slnx
@@ -15,6 +15,7 @@
     <File Path="../Documentation/RevitCommand.md" />
     <File Path="../Documentation/RevitCommandAvailability.md" />
     <File Path="../Documentation/RevitContext.md" />
+    <File Path="../Documentation/RevitContextDesign.md" />
     <File Path="../Documentation/RevitEventHandler.md" />
     <File Path="../Documentation/RevitTask.md" />
     <File Path="../Documentation/RevitUpdater.md" />
